+++ image_classification/CHANGES.rst
@@ -1,7 +1,7 @@
 Changelog
 =========
 
-0.0.10 (2020-??-??)
+0.0.10 (2021-01-25)
 ------------------
 
 - removed ability to split images into grid from `poll.py`